Change the code to use a class
- zemax to speos: working
- speos to zemax: still fixing bugs
