+++ ansys_optical_automation/application/BSDF_converter_example.py
@@ -2,12 +2,7 @@ import os
 import tkinter as tk
 from tkinter import filedialog
 
-from ansys_optical_automation.interop_process.Convert_BSDF_Zemax_to_Speos import (
-    convert_speos_to_zemax_bsdf,
-)
-from ansys_optical_automation.interop_process.Convert_BSDF_Zemax_to_Speos import (
-    convert_zemax_to_speos_bsdf,
-)
+from ansys_optical_automation.interop_process.BSDF_converter import BsdfStructure
 
 
 def getfilename(extension, save=False):
@@ -36,20 +31,23 @@ def getfilename(extension, save=False):
 
 
 def main():
-    """Main script to convert a Zemax/Speos BSDF file to a Speos/Zemax BSDF file"""
+    """Main script to convert BSDF files"""
+
+    bsdf_data = BsdfStructure()
 
     BSDF_inputFilepath = getfilename("*.bsdf *.brdf")
     input_file_extension = os.path.splitext(BSDF_inputFilepath)[1].lower()[0:]
     if not ("bsdf" in input_file_extension or "brdf" in input_file_extension):
-
         msg = "Nonsupported file selected"
         raise TypeError(msg)
 
-    if "bsdf" in input_file_extension:
-        convert_zemax_to_speos_bsdf(BSDF_inputFilepath)
+    bool_log = 1
+    bsdf_data.import_data(BSDF_inputFilepath, bool_log)
+    if bsdf_data.bool_success == 1:
+        if "bsdf" in input_file_extension:
+            bsdf_data.write_speos_file()
 
-    if "brdf" in input_file_extension:
-        convert_speos_to_zemax_bsdf(BSDF_inputFilepath, 1)
-
+        if "brdf" in input_file_extension:
+            bsdf_data.write_zemax_file()
 
 main()
