+++ ansys_optical_automation/post_process/dpf_brdf_viewer.py
@@ -71,7 +71,7 @@ class BrdfStructure:
 
         """
         theta_brdf = [
-            [MeasurePoint.theta, MeasurePoint.brdf]
+            [MeasurePoint.theta_input, MeasurePoint.brdf]
             for MeasurePoint in self.measurement_2d_brdf
             if (MeasurePoint.incidence == incident and MeasurePoint.wavelength == wavelength)
         ]
