+++ tests/workflows/run_test_12_anisotropicbsdf_viewer.py
@@ -1,7 +1,7 @@
 import os
 
-from ansys_optical_automation.interop_process.Convert_BSDF_Zemax_to_Speos import (
-    convert_zemax_to_speos_bsdf,
+from ansys_optical_automation.interop_process.BSDF_converter import (
+    obsolete_convert_zemax_to_speos_bsdf,
 )
 
 
@@ -17,4 +17,4 @@ def unittest_run():
     cwd = os.path.dirname(os.path.realpath(__file__))
     inputFilepath = os.path.join(cwd, "example_models", "test_12_anisotropicbsdf_viewer_zemax.bsdf")
     outputFilepath = os.path.join(cwd, "example_models", "test_12_anisotropicbsdf_viewer_speos.bsdf")
-    convert_zemax_to_speos_bsdf(inputFilepath, outputFilepath, 1, 1)
+    obsolete_convert_zemax_to_speos_bsdf(inputFilepath, outputFilepath, 1, 1)
